#2 refactor: move app logic to app.rs, keep only main fn in main.rs
